Users should be able to exercise their Shield and Underwriter tokens after market expiry
We have a ShieldPayoutDao that handles payouts based on a PayoutGraph.
In this commit, Shield and Underwriter tokens can be exercised only via the Underwriter Pool.
As of now, the implementation seems to be sub-optimal, as it involves an extra contract call between UnderwriterPoolDao and ShieldPayoutDao.
Since this can be improved, this commit does not cover tests for ShieldPayout. A more optimal implementation with a direct call to the SheildPayout will be implemented in a future commit.