Make it possible for contributors to redeem Underwriter Pool Tokens
This commit involves some breaking changes, and tests have not been written to support the changes.
Shield and Underwriter Tokens return payouts between 0 and 1.
When availing a loan, the number of s and i tokens must be equal to the loan amount.
Note : All s tokens are of the same M[D]FT type. The same applies to all i tokens.